requirements.txt
torch==2.1.1
torch_snippets==0.508
albumentations==1.3.1
numpy==1.24.4
Pillow==9.0.1
opencv-python==4.8.1.78
torchvision==0.16.1